feat: extract language server into its own package
