+++ packages/ripple-language-server/bin/ripple-language-server.js
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+
+const { createRippleLanguageServer } = require('../src/server.js');
+
+createRippleLanguageServer();
+
