+++ packages/ripple-language-server/src/diagnosticPlugin.js
@@ -122,11 +122,10 @@ function parseCompilationErrorWithDocument(error, fallbackFileName, sourceText, 
 
 			// Convert 1-based line/column to 0-based for VS Code
 			const zeroBasedLine = Math.max(0, line - 1);
-			const zeroBasedColumn = Math.max(0, column - 1);
+			const actualColumn = Math.max(0, column - 1);
 
 			// Use the original error coordinates from the Ripple compiler
 			// Just use the compiler's position as-is, with a simple 1-character highlight
-			let actualColumn = zeroBasedColumn;
 			let length = Math.min(1, sourceText.split('\n')[zeroBasedLine]?.length - actualColumn || 1);
 
 			return {
@@ -204,3 +203,4 @@ function getEmbeddedInfo(context, document) {
 module.exports = {
 	createRippleDiagnosticPlugin,
 };
+
